Change indexes on build table
Replace `projectid_parentid_starttime` with `projectid_starttime_parentid`.
Order matters in a composite index, and out of these three columns, parentid
is the one that is mostly likely to be left out of the WHERE clause.

This newly reordered composite index allows us to safely remove the separate
indexes we had on the `projectid` and `starttime` columns.